requirements.txt
flet>=0.21.1
edge-tts==6.1.9
pydub==0.25.1
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
googletrans
moviepy
openai-whisper